+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Fixes an issue where OpenGraph links with a size hint would fail downloading. (#32)
 * Improves transformations of starter packs into tibbles with cleaner unnesting. (#31)
+* Allows for more than 25 actors in `bs_get_profile()`, (#29, #30).
 
 # bskyr 0.3.0